mybatis plus

+++ paascloud-common/paascloud-common-core/src/main/java/com/paascloud/core/web/page/TableDataInfo.java
@@ -31,7 +31,7 @@ public class TableDataInfo<T> implements Serializable
      * @param list 列表数据
      * @param total 总记录数
      */
-    public TableDataInfo(List<T> list, int total)
+    public TableDataInfo(List<T> list, long total)
     {
         this.rows = list;
         this.total = total;
